# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church Youth Group Coordination API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  accessCode: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  member: 'member',
  presidency: 'presidency',
  leader: 'leader',
} as const;

export type AuthUserStatus = typeof AuthUserStatus[keyof typeof AuthUserStatus];


export const AuthUserStatus = {
  active: 'active',
  inactive: 'inactive',
  archived: 'archived',
} as const;

export interface AuthUser {
  id: number;
  firstName: string;
  lastName: string;
  role: AuthUserRole;
  groupId: number;
  groupName: string;
  status: AuthUserStatus;
  accessCode: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  member: 'member',
  presidency: 'presidency',
  leader: 'leader',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  inactive: 'inactive',
  archived: 'archived',
} as const;

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  role: UserRole;
  groupId: number;
  groupName: string;
  status: UserStatus;
  accessCode: string;
  createdAt?: string;
}

export type InviteUserInputRole = typeof InviteUserInputRole[keyof typeof InviteUserInputRole];


export const InviteUserInputRole = {
  member: 'member',
  presidency: 'presidency',
  leader: 'leader',
} as const;

export interface InviteUserInput {
  firstName: string;
  lastName: string;
  role?: InviteUserInputRole;
  /** @nullable */
  customAccessCode?: string | null;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  member: 'member',
  presidency: 'presidency',
  leader: 'leader',
} as const;

export type UserUpdateStatus = typeof UserUpdateStatus[keyof typeof UserUpdateStatus];


export const UserUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
  archived: 'archived',
} as const;

export interface UserUpdate {
  role?: UserUpdateRole;
  status?: UserUpdateStatus;
  /** @nullable */
  groupId?: number | null;
}

export interface ResetCodeInput {
  /** @nullable */
  customCode?: string | null;
}

/**
 * @nullable
 */
export type ActivityUserVote = typeof ActivityUserVote[keyof typeof ActivityUserVote] | null;


export const ActivityUserVote = {
  up: 'up',
  down: 'down',
} as const;

export interface Activity {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  equipmentNeeded?: string | null;
  /** @nullable */
  suggestedLocation?: string | null;
  activityType: string;
  groupId: number;
  createdById: number;
  createdByName?: string;
  upvotes: number;
  downvotes: number;
  /** @nullable */
  userVote: ActivityUserVote;
  createdAt?: string;
}

export interface ActivityInput {
  title: string;
  description: string;
  /** @nullable */
  equipmentNeeded?: string | null;
  /** @nullable */
  suggestedLocation?: string | null;
  activityType: string;
}

export interface ActivityUpdate {
  title?: string;
  description?: string;
  /** @nullable */
  equipmentNeeded?: string | null;
  /** @nullable */
  suggestedLocation?: string | null;
  activityType?: string;
}

export type VoteInputVote = typeof VoteInputVote[keyof typeof VoteInputVote];


export const VoteInputVote = {
  up: 'up',
  down: 'down',
  none: 'none',
} as const;

export interface VoteInput {
  vote: VoteInputVote;
}

export interface ScheduledActivity {
  id: number;
  date: string;
  activityId: number;
  activityTitle: string;
  activityType?: string;
  /** @nullable */
  personInChargeId?: number | null;
  /** @nullable */
  personInChargeName?: string | null;
  /** @nullable */
  treatsAssigneeId?: number | null;
  /** @nullable */
  treatsAssigneeName?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  equipment?: string | null;
  /** @nullable */
  notes?: string | null;
  groupId: number;
  createdAt?: string;
}

export interface ScheduledActivityInput {
  date: string;
  activityId: number;
  /** @nullable */
  personInChargeId?: number | null;
  /** @nullable */
  treatsAssigneeId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  equipment?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ScheduledActivityUpdate {
  date?: string;
  activityId?: number;
  /** @nullable */
  personInChargeId?: number | null;
  /** @nullable */
  treatsAssigneeId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  equipment?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Lesson {
  id: number;
  date: string;
  title: string;
  topic: string;
  /** @nullable */
  instructorId?: number | null;
  /** @nullable */
  instructorName?: string | null;
  /** @nullable */
  activityId?: number | null;
  /** @nullable */
  activityTitle?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  groupId: number;
  createdAt?: string;
}

export interface LessonInput {
  date: string;
  title: string;
  topic: string;
  /** @nullable */
  instructorId?: number | null;
  /** @nullable */
  activityId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface LessonUpdate {
  date?: string;
  title?: string;
  topic?: string;
  /** @nullable */
  instructorId?: number | null;
  /** @nullable */
  activityId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface RotationMember {
  userId: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  role?: string | null;
}

export interface SacramentRotation {
  id: number;
  date: string;
  groupId: number;
  members: RotationMember[];
  createdAt?: string;
}

export interface SacramentRotationInput {
  date: string;
  /**
     * @minItems 1
     * @maxItems 3
     */
  memberIds: number[];
}

export interface SacramentRotationUpdate {
  date?: string;
  memberIds?: number[];
}

export interface RandomizationResult {
  members: RotationMember[];
}

export type SubstitutionRequestStatus = typeof SubstitutionRequestStatus[keyof typeof SubstitutionRequestStatus];


export const SubstitutionRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  denied: 'denied',
  resolved: 'resolved',
} as const;

export interface SubstitutionRequest {
  id: number;
  rotationId: number;
  rotationDate?: string;
  requesterId: number;
  requesterName?: string;
  /** @nullable */
  reason?: string | null;
  status: SubstitutionRequestStatus;
  groupId: number;
  createdAt?: string;
}

export interface SubstitutionRequestInput {
  rotationId: number;
  /** @nullable */
  reason?: string | null;
}

export type SubstitutionRequestUpdateStatus = typeof SubstitutionRequestUpdateStatus[keyof typeof SubstitutionRequestUpdateStatus];


export const SubstitutionRequestUpdateStatus = {
  approved: 'approved',
  denied: 'denied',
  resolved: 'resolved',
} as const;

export interface SubstitutionRequestUpdate {
  status: SubstitutionRequestUpdateStatus;
}

export interface MyAssignments {
  activities: number[];
  rotations: number[];
  lessons: number[];
}

export interface DashboardSummary {
  upcomingActivities: ScheduledActivity[];
  upcomingRotations: SacramentRotation[];
  upcomingLessons: Lesson[];
  pendingRequests: number;
  myAssignments: MyAssignments;
}

export type ListActivitiesParams = {
type?: string;
};

export type ListScheduledActivitiesParams = {
from?: string;
to?: string;
};

export type ListLessonsParams = {
from?: string;
to?: string;
};

export type ListSacramentRotationsParams = {
from?: string;
to?: string;
};