# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Church Youth Group Coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid access code
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # USERS / MEMBERS
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities for current group
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

    post:
      operationId: createActivity
      tags: [activities]
      summary: Create an activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{activityId}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get an activity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /activities/{activityId}/vote:
    post:
      operationId: voteActivity
      tags: [votes]
      summary: Upvote or downvote an activity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  # SCHEDULED ACTIVITIES
...
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List Sunday lessons for current group
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"

    post:
      operationId: createLesson
      tags: [lessons]
      summary: Schedule a Sunday lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SACRAMENT ROTATIONS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [accessCode]
      properties:
        accessCode:
          type: string
    AuthUser:
      type: object
      required: [id, firstName, lastName, role, groupId, groupName, status, accessCode]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [member, presidency, leader]
        groupId:
          type: integer
        groupName:
          type: string
        status:
          type: string
          enum: [active, inactive, archived]
        accessCode:
          type: string

    # USERS
...
    Activity:
      type: object
      required: [id, title, description, activityType, groupId, createdById, upvotes, downvotes, userVote]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        equipmentNeeded:
          type: ["string", "null"]
        suggestedLocation:
          type: ["string", "null"]
        activityType:
          type: string
        groupId:
          type: integer
        createdById:
          type: integer
        createdByName:
          type: string
        upvotes:
          type: integer
        downvotes:
          type: integer
        userVote:
          type: ["string", "null"]
          enum: ["up", "down", null]
        createdAt:
          type: string
    ActivityInput:
      type: object
      required: [title, description, activityType]
      properties:
        title:
          type: string
        description:
          type: string
        equipmentNeeded:
          type: ["string", "null"]
        suggestedLocation:
          type: ["string", "null"]
        activityType:
          type: string
    ActivityUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        equipmentNeeded:
          type: ["string", "null"]
        suggestedLocation:
          type: ["string", "null"]
        activityType:
          type: string
    VoteInput:
      type: object
      required: [vote]
      properties:
        vote:
          type: string
          enum: [up, down, none]

    # SCHEDULED ACTIVITIES
...
    Lesson:
      type: object
      required: [id, date, title, topic, groupId]
      properties:
        id:
          type: integer
        date:
          type: string
        title:
          type: string
        topic:
          type: string
        instructorId:
          type: ["integer", "null"]
        instructorName:
          type: ["string", "null"]
        activityId:
          type: ["integer", "null"]
        activityTitle:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        groupId:
          type: integer
        createdAt:
          type: string
    LessonInput:
      type: object
      required: [date, title, topic]
      properties:
        date:
          type: string
        title:
          type: string
        topic:
          type: string
        instructorId:
          type: ["integer", "null"]
        activityId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    LessonUpdate:
      type: object
      properties:
        date:
          type: string
        title:
          type: string
        topic:
          type: string
        instructorId:
          type: ["integer", "null"]
        activityId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # SACRAMENT ROTATIONS
...