# lib/api-zod/src/generated/types/lessonInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church Youth Group Coordination API
 * OpenAPI spec version: 0.1.0
 */

export interface LessonInput {
  date: string;
  title: string;
  topic: string;
  /** @nullable */
  instructorId?: number | null;
  /** @nullable */
  activityId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
}